Pull out util functions to helper dir
