+++ src/core/helpers/check-utils.ts
