+++ src/core/network/client.ts
@@ -5,7 +5,7 @@ import * as http from 'http'
 import * as https from 'https'
 import * as path from 'path'
 import fetch, { RequestInfo, RequestInit } from 'node-fetch'
-import { hasProp, hasPropType } from '../../util'
+import { hasProp, hasPropType } from '../helpers/check-utils'
 import { Opts, Telegram } from 'typegram'
 import MultipartStream from './multipart-stream'
 import { ReadStream } from 'fs'
